requirements.txt
requests
python-dotenv
spinners